SCH-200 # Refactor schematix components structure for floorplan page routing
